Set load_factor and initial capacity
The load_factor will stay as 0.8 at all times, thus when 80% of the
capacity has been reached, the capacity should increase at that time.
The initial capacity has been set to 16 as a baseline for new HashMaps
on the basis of it being decently sized and also a factor of 2.