Create Node class to hold key/value pairs
This is being implemented ahead of the #set method in order to have a
way to hold key/value pairs, as the hashmap bucket is simply an array
and only one value can live in a single index.

Additionally a Node can reference a next_node as a linked list in order
to support multiple information in the same bucket; collisions are not
ideal, but forming a linked list is a way to address it, so
implementation of nodes being a linked list is being supported ahead of
time.