Add #set method for filling buckets with Nodes
This method is perfectly capable of either filling an empty bucket with
a Node or updating the value of an existing Node if the keys match.
However, it does not yet deal with collisions. For now all that happens
in the event of a collision is nothing; the operation does not go
through.