Add collision detection to #set, update logic
The original idea was going to be placing Node objects into buckets if
they were the only piece of data, and convert to a linked list in the
case of collisions. Checking for what object types were in buckets and
if the keys matched or not at the same time proved to overcomplicated
the code, however, so the decision was made to create linked lists from
the start, even if only one piece of data lives in it.

As such, keep in mind that when accessing data in the HashMap, you can't
call on a Node directly anymore, you are interacting with a LinkedList
at all times.