Refactor #remove's logic into LinkedList
A feature was added to LinkedList to handle its own deletion logic by
using #remove_at. This method returns the node that was deleted from the
list, which is also returned in HashMap's #remove method. This way, the
logic for removing a node from a LinkedList is isolated from HashMap.
Now HashMap only cares about telling LinkedList to do this, and only
LinkedList concerns itself with how to do so.