Initialize buckets with LinkedList right away
The original logic had it so a LinkedList was created when a bucket first
receives an entry, and then is appended on collision. However in the end
a bucket ALWAYS contains a LinkedList object regardless of how many
nodes it ends up holding, so the logic was changed so that the buckets
array is created with empty LinkedLists in the first place. This removes
the need to do a bunch of nil checking, and makes #set more lightweight.