Implement algorithm for resizing buckets
The hashmap capacity doubles every time the length exceeds the load
capacity, allowing for more data entries and reducing collision chances.
When it does, the buckets are reinitialized with the new capacity and
the already-existing entries are re-#set into the list. This allows for
some collision cases to potentially be alleviated.