Change how #clear does its job
Instead of just calling initialize, it instantiates a new array. This is
to preserve the buckets size relative to the current capacity, as with
the addition of #resize we don't want to accidentally reset capacity
back down to 16.

It also means that #resize itself can call #clear to help DRY up the
code a bit.